fix: minor cubic-dev-ai suggestions
