+++ packages/oauth-provider/src/utils/index.ts
@@ -157,6 +157,49 @@ export function toAudienceClaim(
 	return audience.length === 1 ? audience.at(0) : audience;
 }
 
+/**
+ * Checks the resource parameter, if provided,
+ * and returns either a valid audience or a tagged validation error.
+ */
+export async function checkResource(
+	ctx: GenericEndpointContext,
+	opts: OAuthOptions<Scope[]>,
+	resource: string | string[] | undefined,
+	scopes: string[],
+) {
+	const normalizedResource = toResourceList(resource);
+	const audience = normalizedResource ? [...normalizedResource] : undefined;
+	if (audience) {
+		// Adds /userinfo to audience
+		const hasOpenId = scopes.includes("openid");
+		const baseUrl = ctx.context.baseURL;
+		const userInfoEndpoint = `${baseUrl}/oauth2/userinfo`;
+		if (hasOpenId && !audience.includes(userInfoEndpoint)) {
+			audience.push(userInfoEndpoint);
+		}
+		// Check valid audiences
+		const filteredValidAudiences = opts.validAudiences?.filter(
+			(aud) => aud.length,
+		);
+		const validAudiences = new Set(
+			filteredValidAudiences?.length ? filteredValidAudiences : [baseUrl],
+		);
+		if (hasOpenId) validAudiences.add(userInfoEndpoint);
+		for (const aud of audience) {
+			if (!validAudiences.has(aud)) {
+				return {
+					success: false,
+					error: "invalid_resource",
+				};
+			}
+		}
+	}
+	return {
+		success: true,
+		audience: toAudienceClaim(audience),
+	};
+}
+
 const cachedTrustedClients = new TTLCache<string, SchemaClient<Scope[]>>();
 
 export async function verifyOAuthQueryParams(
